Fix killing and checking for running processes on target
