+++ tests/integration_test/daemon_startup.py
@@ -13,12 +13,9 @@
 
 import logging
 import time
-import subprocess
 from util import (
     ShellProcess,
-    tcpdump_capture,
-    wait_until_process_exits,
-    get_ps_aux_text,
+    get_running_processes_on_target,
     cleanup,
 )
 
@@ -44,8 +41,10 @@ def test_start_someipd(target):
             someipd_process.get_exit_code(),
         )
         logging.info("someipd output:\n%s", someipd_process.get_output())
+        ps_aux_text = get_running_processes_on_target(target)
+        assert "someipd" in ps_aux_text, ps_aux_text
 
-    ps_aux_text = get_ps_aux_text()
+    ps_aux_text = get_running_processes_on_target(target)
     assert "someipd" not in ps_aux_text, ps_aux_text
 
 
@@ -69,51 +68,8 @@ def test_start_gatewayd(target):
             gatewayd_process.get_exit_code(),
         )
         logging.info("gatewayd output:\n%s", gatewayd_process.get_output())
+        ps_aux_text = get_running_processes_on_target(target)
+        assert "gatewayd" in ps_aux_text, ps_aux_text
 
-    ps_aux_text = get_ps_aux_text()
+    ps_aux_text = get_running_processes_on_target(target)
     assert "gatewayd" not in ps_aux_text, ps_aux_text
-
-
-def test_start_someipd_and_gatewayd(target):
-    cleanup(target)
-    subprocess.run(["ip", "route", "add", "224.0.0.0/4", "dev", "tap0"], check=True)
-
-    with tcpdump_capture("udp port 30490", packet_count=1) as tcpdump_process:
-        with ShellProcess(
-            target,
-            "/someipd",
-            args=[
-                "--configuration",
-                "/mw_someip_config.bin",
-                "--service_instance_manifest",
-                "/someipd_mw_com_config.json",
-            ],
-            env="VSOMEIP_CONFIGURATION=/vsomeip.json",
-        ) as someipd_process:
-            assert someipd_process.is_running(), someipd_process.get_output()
-            with ShellProcess(
-                target,
-                "/gatewayd",
-                args=[
-                    "--configuration",
-                    "/mw_someip_config.bin",
-                    "--service_instance_manifest",
-                    "/gatewayd_mw_com_config.json",
-                ],
-            ) as gatewayd_process:
-                assert gatewayd_process.is_running(), gatewayd_process.get_output()
-                console_output = wait_until_process_exits(tcpdump_process, timeout=10.0)
-                logging.info(
-                    "Final tcpdump to capture SOME/IP-SD traffic...\n" + console_output
-                )
-
-                assert gatewayd_process.is_running(), (
-                    gatewayd_process.get_output(),
-                    "exit code: ",
-                    gatewayd_process.get_exit_code(),
-                )
-                assert someipd_process.is_running(), (
-                    someipd_process.get_output(),
-                    "exit code: ",
-                    someipd_process.get_exit_code(),
-                )
